+++ src/stream/src/executor/chain.rs
@@ -12,15 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::error::Result;
-use risingwave_common::try_match_expand;
-use risingwave_pb::stream_plan;
-use risingwave_pb::stream_plan::stream_node::NodeBody;
-use risingwave_storage::StateStore;
-
-use crate::executor::ExecutorBuilder;
-use crate::executor_v2::{BoxedExecutor, ChainExecutor, Executor, RearrangedChainExecutor};
-use crate::task::{ExecutorParams, LocalStreamManagerCore};
+use super::*;
+use crate::executor_v2::{ChainExecutor, RearrangedChainExecutor};
 
 pub struct ChainExecutorBuilder;
