+++ src/stream/src/executor/global_simple_agg.rs
@@ -14,20 +14,14 @@
 
 //! Streaming Aggregators
 
-use itertools::Itertools;
 use risingwave_common::catalog::TableId;
-use risingwave_common::error::Result;
-use risingwave_common::try_match_expand;
-use risingwave_pb::stream_plan;
-use risingwave_pb::stream_plan::stream_node::NodeBody;
-use risingwave_storage::{Keyspace, StateStore};
 
-use crate::executor::ExecutorBuilder;
+use super::*;
 use crate::executor_v2::aggregation::AggCall;
-use crate::executor_v2::{BoxedExecutor, Executor, SimpleAggExecutor};
-use crate::task::{build_agg_call_from_prost, ExecutorParams, LocalStreamManagerCore};
+use crate::executor_v2::SimpleAggExecutor;
+use crate::task::build_agg_call_from_prost;
 
-pub struct SimpleAggExecutorBuilder {}
+pub struct SimpleAggExecutorBuilder;
 
 impl ExecutorBuilder for SimpleAggExecutorBuilder {
     fn new_boxed_executor(
