+++ src/stream/src/executor/mview.rs
@@ -13,15 +13,10 @@
 // limitations under the License.
 
 use risingwave_common::catalog::{ColumnId, TableId};
-use risingwave_common::try_match_expand;
 use risingwave_common::util::sort_util::OrderPair;
-use risingwave_pb::stream_plan;
-use risingwave_pb::stream_plan::stream_node::NodeBody;
-use risingwave_storage::{Keyspace, StateStore};
 
-use crate::executor::{ExecutorBuilder, Result};
-use crate::executor_v2::{BoxedExecutor, Executor, MaterializeExecutor};
-use crate::task::{ExecutorParams, LocalStreamManagerCore};
+use super::*;
+use crate::executor_v2::MaterializeExecutor;
 
 pub struct MaterializeExecutorBuilder;
