+++ src/stream/src/executor_v2/mod.rs
@@ -50,7 +50,7 @@ mod error;
 mod filter;
 mod global_simple_agg;
 mod hash_agg;
-mod hash_join;
+pub mod hash_join;
 mod hop_window;
 mod local_simple_agg;
 mod lookup;
@@ -83,10 +83,10 @@ pub use filter::FilterExecutor;
 pub use global_simple_agg::SimpleAggExecutor;
 pub use hash_agg::HashAggExecutor;
 pub use hash_join::*;
-pub use hop_window::{HopWindowExecutor, HopWindowExecutorBuilder};
+pub use hop_window::HopWindowExecutor;
 pub use local_simple_agg::LocalSimpleAggExecutor;
 pub use lookup::*;
-pub use lookup_union::{LookupUnionExecutor, LookupUnionExecutorBuilder};
+pub use lookup_union::LookupUnionExecutor;
 pub use merge::MergeExecutor;
 pub use mview::*;
 pub use project::ProjectExecutor;
@@ -95,7 +95,7 @@ use simple::{SimpleExecutor, SimpleExecutorWrapper};
 pub use source::*;
 pub use top_n::TopNExecutor;
 pub use top_n_appendonly::AppendOnlyTopNExecutor;
-pub use union::{UnionExecutor, UnionExecutorBuilder};
+pub use union::UnionExecutor;
 
 pub type BoxedExecutor = Box<dyn Executor>;
 pub type BoxedMessageStream = BoxStream<'static, StreamExecutorResult<Message>>;
