+++ src/stream/src/executor_v2/source.rs
@@ -20,22 +20,17 @@ use futures::{Stream, StreamExt};
 use futures_async_stream::try_stream;
 use risingwave_common::array::column::Column;
 use risingwave_common::array::{ArrayBuilder, ArrayImpl, I64ArrayBuilder, StreamChunk};
-use risingwave_common::catalog::{ColumnId, Field, Schema, TableId};
+use risingwave_common::catalog::{ColumnId, Schema, TableId};
 use risingwave_common::error::ErrorCode::InternalError;
-use risingwave_common::error::{Result, RwError, ToRwResult};
-use risingwave_common::try_match_expand;
+use risingwave_common::error::{Result, RwError};
 use risingwave_connector::SplitImpl;
-use risingwave_pb::stream_plan;
-use risingwave_pb::stream_plan::stream_node::NodeBody;
 use risingwave_source::*;
 use risingwave_storage::{Keyspace, StateStore};
-use tokio::sync::mpsc::{unbounded_channel, UnboundedReceiver};
+use tokio::sync::mpsc::UnboundedReceiver;
 
 use super::error::StreamExecutorError;
 use super::monitor::StreamingMetrics;
 use super::*;
-use crate::executor::ExecutorBuilder;
-use crate::task::{ExecutorParams, LocalStreamManagerCore};
 
 /// [`SourceExecutor`] is a streaming source, from risingwave's batch table, or external systems
 /// such as Kafka.
@@ -60,69 +55,6 @@ pub struct SourceExecutor {
     stream_source_splits: Vec<SplitImpl>,
 
     source_identify: String,
-}
-
-pub struct SourceExecutorBuilder {}
-
-impl ExecutorBuilder for SourceExecutorBuilder {
-    fn new_boxed_executor(
-        params: ExecutorParams,
-        node: &stream_plan::StreamNode,
-        store: impl StateStore,
-        stream: &mut LocalStreamManagerCore,
-    ) -> Result<BoxedExecutor> {
-        let node = try_match_expand!(node.get_node_body().unwrap(), NodeBody::Source)?;
-        let (sender, barrier_receiver) = unbounded_channel();
-        stream
-            .context
-            .lock_barrier_manager()
-            .register_sender(params.actor_id, sender);
-
-        let source_id = TableId::from(&node.table_ref_id);
-        let source_desc = params.env.source_manager().get_source(&source_id)?;
-
-        let stream_source_splits = match &node.stream_source_state {
-            Some(splits) => splits
-                .stream_source_splits
-                .iter()
-                .map(|split| SplitImpl::restore_from_bytes(splits.get_split_type().clone(), split))
-                .collect::<anyhow::Result<Vec<SplitImpl>>>()
-                .to_rw_result(),
-            _ => Ok(vec![]),
-        }?;
-
-        let column_ids: Vec<_> = node
-            .get_column_ids()
-            .iter()
-            .map(|i| ColumnId::from(*i))
-            .collect();
-        let mut fields = Vec::with_capacity(column_ids.len());
-        fields.extend(column_ids.iter().map(|column_id| {
-            let column_desc = source_desc
-                .columns
-                .iter()
-                .find(|c| &c.column_id == column_id)
-                .unwrap();
-            Field::with_name(column_desc.data_type.clone(), column_desc.name.clone())
-        }));
-        let schema = Schema::new(fields);
-        let keyspace = Keyspace::executor_root(store, params.executor_id);
-
-        Ok(Box::new(SourceExecutor::new(
-            source_id,
-            source_desc,
-            keyspace,
-            column_ids,
-            schema,
-            params.pk_indices,
-            barrier_receiver,
-            params.executor_id,
-            params.operator_id,
-            params.op_info,
-            params.executor_stats,
-            stream_source_splits,
-        )?))
-    }
 }
 
 impl SourceExecutor {
